\name{parse.ref}
\alias{parse.ref}
\title{Parse either srcrefs, prerefs or pairs of the same.}
\usage{parse.ref(ref, ...)}
\description{Parse either srcrefs, prerefs or pairs of the same.}
\value{List containing the parsed srcref/preref}
\arguments{\item{ref}{the srcref, preref or pair of the same}
\item{\dots}{ignored}}
