\name{parse.toggle}
\alias{parse.toggle}
\title{Turn a binary element on; parameters are ignored.}
\usage{parse.toggle(key, rest)}
\description{Turn a binary element on; parameters are ignored.}
\value{A list with the key and \code{TRUE}}
\arguments{\item{key}{parsing key}
\item{rest}{the expression to be parsed}}
