\name{preorder.walk.expression}
\alias{preorder.walk.expression}
\title{Recursively walk an expression (as returned by parse) in...}
\usage{preorder.walk.expression(proc, expression)}
\description{Recursively walk an expression (as returned by \code{parse}) in
preorder.}
\value{NULL}
\arguments{\item{proc}{the procedure to apply to each subexpression}
\item{expression}{the root of the expression}}
